% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generate_n_names2}
\alias{generate_n_names2}
\title{Generate N names}
\usage{
generate_n_names2(n = 1, Year = 1950, Sex = "F")
}
\description{
Calls \code{generate_name2()} \code{n} times.
}
\details{
See \code{generate_name()}

The first time using \code{most_prob_name2()}, \code{generate_name2()} or
\code{generate_n_names2()} with a unique Year and Sex combination may take
a few seconds to run, but should be faster on any subsequent call to the same
function or any of the other two using the same Year and Sex combination.
}

