% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generate_name}
\alias{generate_name}
\title{Generate One Name}
\usage{
generate_name(Year = 1950, Sex = "F")
}
\description{
Generates one name on the basis of transition probabilities from a given year
for a given sex.
}
\details{
Using data from the \code{babynames} package, this function will generate
novel "names" using a Markov Chain based on transition probabilities for
a given year and sex.

Since the Markov Chains rarely land in the \code{end} state, this function
also probabilistically ends the chain on the basis of how long the name
currently is, based on the distribution of name lengths for the
given year and sex.

The first time using \code{most_prob_name()}, \code{generate_name()} or
\code{generate_n_names()} with a unique Year and Sex combination may take
a few seconds to run, but should be faster on any subsequent call to the same
function or any of the other two using the same Year and Sex combination.
}

