% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generate_name2}
\alias{generate_name2}
\title{Generate One Name2}
\usage{
generate_name2(Year = 1950, Sex = "F")
}
\description{
Generates one name on the basis of transition probabilities from bigrams to the next character
from a given year for a given sex.
}
\details{
Using data from the \code{babynames} package, this function will generate
novel "names" using a Markov Chain based on transition probabilities from bigrams to the next character
for a given year and sex.
When the current bigram wasn't present in the training data, it falls back to bigram transition probabilities
(i.e. character -> character probabilities).

This function has a maximum name length of 20 characters.

The first time using \code{most_prob_name2()}, \code{generate_name2()} or
\code{generate_n_names2()} with a unique Year and Sex combination may take
a few seconds to run, but should be faster on any subsequent call to the same
function or any of the other two using the same Year and Sex combination.
}

