% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generate_states}
\alias{generate_states}
\title{Generate States}
\usage{
generate_states(Year = 1950, Sex = "F")
}
\arguments{
\item{Year}{the year from which training data is drawn}

\item{Sex}{the sex from which training data is drawn}
}
\description{
Generates the transition matrix from one character
to the next for a given year and sex.
}

