% Generated by roxygen2 (4.0.2): do not edit by hand
\name{most_prob_name}
\alias{most_prob_name}
\title{Most Probable Path}
\usage{
most_prob_name(Year = 1950, Sex = "F")
}
\arguments{
\item{Year}{the year from which training data is drawn}

\item{Sex}{the sex from which training data is drawn}
}
\description{
Returns the most probable path through the character \eqn{\rightarrow}{->} character state space for
a given year for a given sex.
}
\details{
The most probable path for most years and sexes is occasionally an infinite
loop. When a path is about to loop infinitely, the path stops, and \code{:}
is appended to the end. If it reaches the end state, \code{.} is appended.

The first time using \code{most_prob_name()}, \code{generate_name()} or
\code{generate_n_names()} with a unique Year and Sex combination may take
a few seconds to run, but should be faster on any subsequent call to the same
function or any of the other two using the same Year and Sex combination.
}
\examples{
most_prob_name(Year = 1985, Sex = "M")
}

