% Generated by roxygen2 (4.0.2): do not edit by hand
\name{most_prob_name2}
\alias{most_prob_name2}
\title{Most Probable Path2}
\usage{
most_prob_name2(Year = 1950, Sex = "F")
}
\description{
Returns the most probable path through the bigram \eqn{\rightarrow}{→} character state space for
a given year for a given sex.
}
\details{
The most probable path for most years and sexes is occasionally an infinite
loop. When a path is about to loop infinitely, the path stops, and \code{:}
is appended to the end. If it reaches the end state, \code{.} is appended.

The first time using \code{most_prob_name2()}, \code{generate_name2()} or
\code{generate_n_names2()} with a unique Year and Sex combination may take
a few seconds to run, but should be faster on any subsequent call to the same
function or any of the other two using the same Year and Sex combination.
}
\examples{
most_prob_name(Year = 1985, Sex = "M")
}

